+++ src/main/java/com/aetherteam/aether/item/AetherItems.java
@@ -95,9 +95,9 @@ public class AetherItems {
     public static final Component GOLD_DUNGEON_TOOLTIP = Component.translatable("aether.dungeon.gold_dungeon").withStyle(Style.EMPTY.withItalic(true).withColor(TextColor.parseColor("#FDF55F").result().get()));
 
     // Tools
-    public static final PickaxeItem SKYROOT_PICKAXE = register("skyroot_pickaxe", SkyrootPickaxeItem::new);
-    public static final AxeItem SKYROOT_AXE = register("skyroot_axe", SkyrootAxeItem::new);
-    public static final ShovelItem SKYROOT_SHOVEL = register("skyroot_shovel", SkyrootShovelItem::new);
+    public static final PickaxeItem SKYROOT_PICKAXE = register("skyroot_pickaxe", properties -> new SkyrootPickaxeItem(properties));
+    public static final AxeItem SKYROOT_AXE = register("skyroot_axe", properties -> new SkyrootAxeItem(properties));
+    public static final ShovelItem SKYROOT_SHOVEL = register("skyroot_shovel", properties -> new SkyrootShovelItem(properties));
     public static final HoeItem SKYROOT_HOE = register("skyroot_hoe", SkyrootHoeItem::new);
 
     public static final PickaxeItem HOLYSTONE_PICKAXE = register("holystone_pickaxe", HolystonePickaxeItem::new);