+++ src/main/java/com/aetherteam/aether/item/tools/skyroot/SkyrootAxeItem.java
@@ -9,7 +9,7 @@ import net.minecraft.world.item.Item;
  * Double drops behavior is called by {@link com.aetherteam.aether.loot.functions.DoubleDrops}.
  */
 public class SkyrootAxeItem extends AxeItem implements SkyrootTool {
-    public SkyrootAxeItem() {
-        super(AetherItemTiers.SKYROOT, 6.0F, -3.2F, new Item.Properties());
+    public SkyrootAxeItem(Item.Properties properties) {
+        super(AetherItemTiers.SKYROOT, 6.0F, -3.2F, properties);
     }
 }